backend/requirements.txt
asgiref
channels
channels_redis
daphne
dj-database-url
django
django-heroku
djangorestframework
Pillow
pyyaml
uritemplate
django-rest-auth[with_social]
psycopg2
whitenoise
